pyproject.toml
[project]
name = "tts-dataset-maker"
version = "1.0.0"
description = "Simple local TTS dataset creation pipeline"
authors = [{name = "TTS Dataset Maker", email = "team@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "assemblyai>=0.17.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "label-studio-sdk>=1.0.16",
    "torch==2.1.0",
    "torchaudio==2.1.0",
    "silero-vad>=4.0.0",
    "deepfilternet>=0.5.0",
    "pydub>=0.25.1",
    "label-studio>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tts_pipeline"]
